Backport "Make sure that when 32-bit connection-IDs roll over, still in-use IDs are not re-issued"
Summary:
This is a backport of
https://github.com/mysql/mysql-server/commit/cdb6ddc9c35bba01d9d3a20cf25eac570b5e5bf3

There are two problems fixed in this change. The first is that the
connection_id field in the initial handshake is only 32 bits, but the server
stores it as 64 bits. This means that as soon as the ids exceed UINT32_MAX,
the data in it becomes meaningless. This is fixed by having the server store
connection id as 32 bits as well. The second problem is that once connection
ids wrap around, it is possible to reuse ids currently in use. The fix is to
store a std::set of ids in use, and check if an id is already in use before
using it.

There are a few differences between this change, and the change from upstream.
- I didn't bother porting the whole Global_THD_manager class over.
- I'm using std::set instead of an array. It seems like a set should be more efficient.
- I'm reusing the LOCK_thread_count to protect the std::set instead of creating a new lock.
- There were places where ulong was changed to my_thread_id that I didn't bother porting. The idea is that this will be impossible to track during rebases anyway. As long as the thread_id assignment logic is fixed, I think we should be relatively safe during rebases, since we'd just reading a 32 bit integer as a 64 bit one. I've also changed thread_id -> thread_id_counter, so any attempt to modify that variable should become a compilation failure.
- For some reason, the "Connections" status variable in upstream was essentially set to thread_id_counter. This means that it may not be monotonically increasing any more. With a 64 bit number, this was probably never an issue, but with 32 bits, wrap around is a real possibility. To preserve current behaviour, I've added the `total_thread_ids` counter which is a 64 bit integer tracking how many connections were made in total.

Test Plan: Wrote new test main.unique_thread_id

Reviewers: jkedgar

Reviewed By: jkedgar

Subscribers: anca, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55899
Differential Revision: https://reviews.facebook.net/D55971
Differential Revision: https://reviews.facebook.net/D56301

+++ sql/filesort.cc
@@ -461,11 +461,11 @@ ha_rows filesort(THD *thd, TABLE *table, Filesort *filesort,
 
     if (log_warnings > 1)
     {
-      sql_print_warning("%s, host: %s, user: %s, thread: %lu, query: %-.4096s",
+      sql_print_warning("%s, host: %s, user: %s, thread: %u, query: %-.4096s",
                         ER_THD(thd, ER_FILSORT_ABORT),
                         thd->security_ctx->host_or_ip,
                         &thd->security_ctx->priv_user[0],
-                        (ulong) thd->thread_id,
+                        thd->thread_id(),
                         thd->query());
     }
   }
